# api/openapi.yml
openapi: '3.0.3'
info:
  title: 'Zone42 API'
  description: 'Zone42 administrator api'
  version: '1.0'
servers:
  - url: https://api.zone-42.com
security:
  - bearerAuth: []
paths:
  /zones:
    get:
      summary: 'list of all zones'
      parameters:
        - $ref: '#/components/parameters/start'
        - $ref: '#/components/parameters/count'
        - $ref: '#/components/parameters/ascending'
        - $ref: '#/components/parameters/q'
      responses:
        200:
          $ref: '#/components/responses/list_response'
        400:
          $ref: '#/components/responses/error_response'

    post:
      summary: 'create a new zone'
      requestBody:
        $ref: '#/components/requestBodies/new_zone'
      responses:
        201:
          $ref: '#/components/responses/success_response'
        400:
          $ref: '#/components/responses/error_response'
        401:
          $ref: '#/components/responses/error_response'
        409:
          $ref: '#/components/responses/error_response'
...
components:
  parameters:
    start:
      name: start
      in: query
      description: 'start index'
      required: false
      schema:
        type: integer
    count:
      name: count
      in: query
      description: 'number of items in response'
      required: false
      schema:
        type: integer
    ascending:
      name: ascending
      in: query
      description: 'ascending order'
      required: false
      schema:
        type: boolean
    q:
      name: q
      in: query
      description: 'search query'
      required: false
      schema:
        type: string
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  requestBodies:
    new_zone:
      description: 'zone'
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/new_zone'
...
  responses:
    list_response:
      description: 'list response'
      content:
        application/json:
          schema:
            title: list response
            type: object
            properties:
              total:
                type: integer
              items:
                type: array
                items:
                  type: object
                  properties:
                    id:
                      type: string
                    enabled:
                      type: boolean
    error_response:
      description: 'error response'
      content:
        application/json:
          schema:
            title: error response
            type: object
            properties:
              code:
                type: integer
              message:
                type: string
    success_response:
      description: 'successful response'
      content:
        application/json:
          schema:
            title: successful response
            type: object
            properties:
              code:
                type: integer
              message:
                type: string
              "data":
                type: object
  schemas:
...
    new_zone:
      title: new zone
      type: object
      properties:
        name:
          type: string
        enabled:
          type: boolean
        dnssec:
          type: boolean
        cname_flattening:
          type: boolean
      example: '{"name": "example.com.", "enabled": true, "dnssec":false, "cname_flattening": false}'
...